Module Awaitable.Awaiter

Ability to await for a signal from the past.

Awaitable.await only receives a signal at or after the point of calling it. This API allows the awaiting process to be broken into two steps, add and await, such that a signal after add can be received by await.

type 'a awaitable := 'a t

An erased type alias for Awaitable.t.

type t

Represents a single use awaiter of a signal to an awaitable.

val add : 'a awaitable -> t

add awaitable create a single use awaiter, adds it to the FIFO associated with the awaitable, and returns the awaiter.

val await : t -> unit

await awaiter awaits for the association awaitable to be signaled.

val remove : t -> unit

remove awaiter marks the awaiter as having been signaled and removes it from the FIFO associated with the awaitable.

⚠️ An explicit call of remove is needed when an added awaiter is not awaited for. In such a case, from the point of view of lost signals, the caller of remove should be considered to have received or consumed a signal before the call of remove.