Module Picos_std_sync.Semaphore

Counting and Binary semaphores.

ℹ️ This intentionally mimics the interface of Stdlib.Semaphore. Unlike with the standard library semaphores, blocking on these semaphores allows an effects based scheduler to run other fibers on the thread.

module Counting : sig ... end

A counting semaphore.

module Binary : sig ... end

A binary semaphore.